requirements.txt
ics==0.7.2
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
lxml==5.3.0
pytz==2024.1
